Clean up some tests
